+++ src/Config/Settings.ts
@@ -1,9 +1,11 @@
 import { makeDefaultSuggestionBuilder } from '../Suggestor/Suggestor';
-import { DATAVIEW_SYMBOLS, DEFAULT_SYMBOLS } from '../TaskSerializer/DefaultTaskSerializer';
+import { DEFAULT_SYMBOLS } from '../TaskSerializer/DefaultTaskSerializer';
+import { DATAVIEW_SYMBOLS } from '../TaskSerializer/DataviewTaskSerializer';
 import { StatusConfiguration } from '../StatusConfiguration';
 import { Status } from '../Status';
 import { DefaultTaskSerializer, type TaskSerializer } from '../TaskSerializer';
 import type { SuggestionBuilder } from '../Suggestor';
+import { DataviewTaskSerializer } from '../TaskSerializer/DataviewTaskSerializer';
 import { DebugSettings } from './DebugSettings';
 import { StatusSettings } from './StatusSettings';
 import { Feature } from './Feature';
@@ -40,7 +42,7 @@ export const TASK_FORMATS = {
     },
     dataview: {
         displayName: 'Dataview',
-        taskSerializer: new DefaultTaskSerializer(DATAVIEW_SYMBOLS),
+        taskSerializer: new DataviewTaskSerializer(),
         buildSuggestions: makeDefaultSuggestionBuilder(DATAVIEW_SYMBOLS),
     },
 } as const;
